super basic database
This commit has a basic connection to the db. The list of students is
populated by the database now. This method only updates the list of
students on list page instantiation. This method should work for now
given that a new student is almost certainly not added while someone is
checking that student in but if we want it to be a live updating thing
it might be possible.