Users db added and small update
The users db has been added, the only service the userprovider gives
right now is to get a user by input id. I have also updated
classroom-id.ts to use this service and check the user permission.

I added a side note to list.html that shows one way that we can do live
updating for students information but I'm not certain of the drawbacks
so it might need further experimenting.