Data Models added
Added basic data models for student, user, and classroom.
All are under models/db-models so that if we need to import multiple
models in a single page we don't clutter the imports with multiple
lines.
Also did a little bit of code review to get rid of some code bloat.