making list better
List no longer stores the student information. I realized we were
instantiating multiple list objects, one for each tab, and this meant
duplicating the student data several times.
Instead the students are just stored in the student provider and
accessed in list.
I also did this because it made the synchronization better and
eventually we would need this anyway.
I also added a map to Iterable pipe to make accessing the map values
easier.