Added a loading screen
This loading screen should solve the initial problems of timing and data
base connections. We may want to review the styling and other design
choices related to it but funcitonally it works as intended. I also
added a toast to prevent the error where a user clicks on the classroom
drop down but clicks ok without selecting a room.