login page changes
Updated the login page so that the classroom selection is more
consistant with the other room type buttons. Currently I'm using every
room as a button but I foresee this being very ugly with more than 4
rooms. A solution to the ugliness would be to make it a set of radio
buttons instead, and this fix should be straightforward.