admin student modal updates
Added the functionality to allow an admin to provide an ID when creating
new students, rather than just taking the next available.

Currently this has some UI confusion in that when adding a new student
the provided ID must be either a positive number that isn't already
taken, or -1. I'm not sure which solution we want to pursue but as is I
think this will be confusing to the admins.