Admin reporting page updates
Added the ability to change the week that the user gets billing for. It
takes any specific day as the field, but then gets information for the
whole week, i.e. user choosing tuesday 8th of january is the same as the
user choosing wednesday 9th of january. Currently if the user selects a
week for which we have no billing data then the tables will display text
"there is no billing information for this week".