admin reporting page updates
Added a loop to set all of the billable hours to 0 if they were
negative. This way the exported csv does not have -1 for a person with
no billable hours. I did this in the setup method, but if we need to it
could also be done inside the toCSV method so that the information isn't
"lost" (negative numbers set to 0), until it finally tries to export.