reporting details page updates
Added the ability to edit time in and out
there is an ugly edge case I  may need to address if a user ever clears
time out resulting in thousands of negative hours which could possibly
make its way to billing.
this can never happen automatically, only if the admin clears the time
and doesn't fix it, so hopefully an admin would notice the very big
number in the hours fields and act to change it.