auth and constant updates
to get the project in a more stable state before I do a longer revision
of some code I made it so that auth and constants providers, which will
be removed and incorporated into different db's, reference the google
cloud couchdb instance, whereas all the other providers reference
cloudant. this effectively fixes the "too many connections" issue but in
the future we don't want to rely on two different DBaaS providers.