auth provider, constants provider updates
Mainly changes to auth provider, setup so that there is no local db,
this might be the more secure way to do it anyway, even though
everything is hashed and it shouldn't be a problem it will at least not
load the whole table into the users browser. Also might fix the number
of connections issue without needing to change the code too much.