provider updates
The code should be setup where the user has to enter in the db
credentials first, this should prevent the developer tools snooping that
we were afraid of. It should also have a solution for "remembering" if
the user has logged on before so that they don't have to enter the
credentials everytime. The current implementation is not robust so
initial.ts and utility provider could use a code review.